copied files
I copied the python scripts I had in my scripts repository, because I need many of them for use with other things in my devkit. I will hold the python scripts in my scripts repo as well, as I actively use both in different contexts, on different machines, etc.
Also wrote a new class and an algorithm to go with it.